Document of Tools for Pentesting, OSINT and Hacking
The repository contains comprehensive documentation on various phases of penetration testing, along with details of tools and techniques used in each phase. This repository aims to serve as a guide for aspiring security professionals and a reference for experienced ones. It covers topics such as reconnaissance, enumeration, vulnerability scanning, exploitation, post-exploitation, and reporting. Additionally, it includes links to open-source and commercial tools commonly used in penetration testing. The documentation is regularly updated to keep pace with the latest advancements in the field.